+++ PDColorPicker/Classes/UIView+Layout.swift
@@ -11,8 +11,39 @@ import UIKit
 @available(iOS 9.0, *)
 extension UIView {
 
-  // MARK: - Main Anchoring
+  /**
+   Provides a convenient method for anchoring the receiver to other
+   anchors within its view hierarchy. All parameters have default values of
+   `nil` or `0` so only the desired constraints need be provided.
 
+   This method should only be called *after* the receiver has been added to its superview,
+   otherwise an exception may be thrown.
+
+   This method sets `translatesAutoresizingMaskIntoConstraints` to `false` on
+   the receiver.
+
+   Example:
+
+   ```
+   // in a UIView subclass:
+   let aView = UIView()
+   addSubview(aView)
+   // sets aView to be positioned 10 points from the
+   //   superview's left anchor
+   aView.anchor(left: viewAnchor, leftConstant: 10))
+   ```
+
+   - parameter left: The anchor to pin the left side of the receiver to.
+   - parameter right: The anchor to pin the right side of the receiver to.
+   - parameter top: The anchor to pin the top side of the receiver to.
+   - parameter bottom: The anchor to pin the bottom side of the receiver to.
+   - parameter leftConstant: The number of points of spacing given to the left side (increases to the right).
+   - parameter rightConstant: The number of points of spacing given to the right side (increases to the left).
+   - parameter topConstant: The number of points of spacing given to the top side (increases down).
+   - parameter bottomConstant: The number of points of spacing given to the bottom side (increases up).
+   - parameter height: The fixed height in points.
+   - parameter width: The fixed width in points.
+  */
   func anchor(left: NSLayoutXAxisAnchor? = nil,
               right: NSLayoutXAxisAnchor? = nil,
               top: NSLayoutYAxisAnchor? = nil,
@@ -22,9 +53,7 @@ extension UIView {
               topConstant: CGFloat = 0,
               bottomConstant: CGFloat = 0,
               height: CGFloat = 0,
-              width: CGFloat = 0,
-              heightMult: CGFloat? = nil,
-              widthMult: CGFloat? = nil) {
+              width: CGFloat = 0) {
 
     translatesAutoresizingMaskIntoConstraints = false
 
@@ -51,90 +80,6 @@ extension UIView {
     if height > 0 {
       heightAnchor.constraint(equalToConstant: height).isActive = true
     }
-
-    if let heightMult = heightMult, let superview = superview {
-      anchorHeightProportionally(to: superview.heightAnchor, multiplier: heightMult)
-    }
-
-    if let widthMult = widthMult, let superview = superview {
-      anchorWidthProportionally(to: superview.widthAnchor, multiplier: widthMult)
-    }
-  }
-
-  // MARK: - Centering
-
-  func anchorCenter(in view: UIView? = nil, padding: CGFloat? = nil) {
-    guard let centerView = view ?? superview else { superviewError() }
-    
-    if let padding = padding {
-      anchor(
-        left: centerView.leftAnchor,
-        right: centerView.rightAnchor,
-        top: centerView.topAnchor,
-        bottom: centerView.bottomAnchor,
-        leftConstant: padding,
-        rightConstant: padding,
-        topConstant: padding,
-        bottomConstant: padding
-      )
-    }
-
-    anchorCenterHorizontally(in: view)
-    anchorCenterVertically(in: view)
-  }
-
-  func anchorCenterHorizontally(in view: UIView? = nil) {
-    guard let view = view ?? superview else { superviewError() }
-
-    translatesAutoresizingMaskIntoConstraints = false
-    centerXAnchor.constraint(equalTo: view.centerXAnchor).isActive = true
-  }
-
-  func anchorCenterVertically(in view: UIView? = nil) {
-    guard let view = view ?? superview else { superviewError() }
-
-    translatesAutoresizingMaskIntoConstraints = false
-    centerYAnchor.constraint(equalTo: view.centerYAnchor).isActive = true
-  }
-
-  // MARK: - Proportional Sizing
-
-  func anchorHeightProportionally(to dimension: NSLayoutDimension, multiplier: CGFloat) {
-    translatesAutoresizingMaskIntoConstraints = false
-    heightAnchor.constraint(equalTo: dimension, multiplier: multiplier).isActive = true
-  }
-
-  func anchorWidthProportionally(to dimension: NSLayoutDimension, multiplier: CGFloat) {
-    translatesAutoresizingMaskIntoConstraints = false
-    widthAnchor.constraint(equalTo: dimension, multiplier: multiplier).isActive = true
-  }
-
-  // MARK: - Fill
-
-  func anchorFill(in view: UIView? = nil, padding: CGFloat = 0) {
-    guard let fillView = view ?? superview else { superviewError() }
-
-    let paddingInsets = UIEdgeInsets(top: padding, left: padding, bottom: padding, right: padding)
-    anchorFill(in: fillView, paddingInsets: paddingInsets)
-  }
-
-  func anchorFill(in view: UIView, paddingInsets: UIEdgeInsets) {
-    anchor(
-      left: view.leftAnchor,
-      right: view.rightAnchor,
-      top: view.topAnchor,
-      bottom: view.bottomAnchor,
-      leftConstant: paddingInsets.left,
-      rightConstant: paddingInsets.right,
-      topConstant: paddingInsets.top,
-      bottomConstant: paddingInsets.bottom
-    )
-  }
-
-  // MARK: - Error Handling
-
-  private func superviewError() -> Never {
-    fatalError("Could not find a superview for view:\n\n\(self.debugDescription)\n\nThe view may not have been added to a superview")
   }
 
 }