+++ CHANGELOG.md
@@ -1,9 +1,16 @@
+# [PDColorPicker v0.4.0](https://github.com/pdil/PDColorPicker/releases/tag/0.4.0)
+Released on Thursday, August 31, 2017.
+
+### Improvements
+* Removed unnecessary layout anchor functions.
+* Improved framework documentation.
+
 # [PDColorPicker v0.3.0](https://github.com/pdil/PDColorPicker/releases/tag/0.3.0)
 Released Monday, August 28, 2017.
 
 ### Improvements
 * Fixed sizing of `PDColorPickerViewController` for different screen sizes (e.g. iPad in multiple split-view sizes)
-* Add the following named colors to `PDColor` for convenience:
+* Added the following named colors to `PDColor` for convenience:
   * `PDColor.red`
   * `PDColor.orange`
   * `PDColor.yellow`